==> domains/snopes_crawler/00_extract_articles_from_snopes.ipynb <==
from bs4 import BeautifulSoup
import requests
import re
from time import sleep
import io
import tldextract
import glob
from collections import Counter, defaultdict
import pandas as pd
import numpy as np

def get_all_articles_from_page(page,article_div_selector={"class": "article-links-wrapper"}):
    articles = []
    page_counter = 1
    total_pages = 9999
    page = page.rstrip("/")
    while page_counter <= total_pages:
        print 'page: ', page_counter, total_pages
        page_html = requests.get("/".join([page, "page", str(page_counter)]))
        soup = BeautifulSoup(page_html.text)
        total_pages = int(re.sub("Page [0-9]+ of ", "", soup.find("span",{"class":"page-count"}).get_text()))                 
        article_list = soup.find("div", article_div_selector)
        for article in article_list.select("article"):
            #print article
            for item in article.find_all("meta",{"itemprop":"url"}):
                if (item['content'] != "http://www.snopes.com" 
                    and not item['content'][-3:] in ['png','jpg']
                    and not item['content'][-4:] == 'jpeg'):
                    articles.append(item['content'])
        page_counter += 1
        sleep(2)
    return articles
        
def extract_articles(articles):
    for i, article in enumerate(articles):
        if i % 100 == 0:
            print i
        try:
            page_html = requests.get(article)
            of = io.open("snopes_data/article_htmls/"+article.replace("http://www.snopes.com/","").replace("/",""),"w")
            of.write(page_html.text)
            of.close()
        except:
            failed_articles.append(article)
        sleep(1)
    return 


election_articles = get_all_articles_from_page("http://www.snopes.com/tag/election-2016")

facts_articles = get_all_articles_from_page("http://www.snopes.com/category/facts")

f = open("snopes_data/all_article_urls.txt","w")
for article in facts_articles+election_articles:
    f.write(article+"\n")
f.close()

extract_articles(election_articles)

extract_articles(facts_articles)
